backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
Django==5.2.13
django-cors-headers==4.9.0
djangorestframework==3.17.1
docxtpl==0.20.2
et_xmlfile==2.0.0
idna==3.11
Jinja2==3.1.6
lxml==6.1.0
MarkupSafe==3.0.3
openpyxl==3.1.5
pycparser==3.0
PyMySQL==1.1.2
python-docx==1.2.0
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
